pyproject.toml
[project]
name = "video-conference-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "channels-redis>=4.3.0",
    "channels[daphne]>=4.3.2",
    "django>=6.0.1",
    "livekit-api>=1.1.0",
    "redis>=7.1.0",
]
